+++ src/backend/scientific_defaults/ScientificDefaults.cpp
@@ -1,5 +1,5 @@
 #include "openPMD/backend/scientific_defaults/ScientificDefaults.hpp"
-#include "openPMD/backend/scientific_defaults/ScientificDefaults_impl.hpp"
+#include "openPMD/backend/scientific_defaults/ConfigAttribute.hpp"
 
 #include "openPMD/IO/AbstractIOHandler.hpp"
 #include "openPMD/backend/Writable.hpp"
